# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductGender = typeof ProductGender[keyof typeof ProductGender];


export const ProductGender = {
  men: 'men',
  women: 'women',
  unisex: 'unisex',
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  gender: ProductGender;
  description: string;
  scentNotes: string[];
  category: string;
  priceCents: number;
  imageUrl: string;
  stock: number;
  featured: boolean;
  createdAt: string;
}

export interface CartItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CartItemUpdate {
  /** @minimum 1 */
  quantity: number;
}

export interface CartItem {
  id: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  cartId: string;
  items: CartItem[];
  subtotalCents: number;
}

export interface OrderInput {
  cartId: string;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  city: string;
  postalCode: string;
}

export interface OrderItem {
  id: number;
  quantity: number;
  priceCentsAtPurchase: number;
  product: Product;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  shippingAddress: string;
  city: string;
  postalCode: string;
  totalCents: number;
  status: string;
  items: OrderItem[];
  createdAt: string;
}

export type ListProductsParams = {
gender?: ProductGender;
category?: string;
search?: string;
};