# lib/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Cart {
  cartId: string;
  items: CartItem[];
  subtotalCents: number;
}