# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  id: number;
  quantity: number;
  product: Product;
}