# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}