# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductGender } from './productGender';

export interface Product {
  id: number;
  name: string;
  brand: string;
  gender: ProductGender;
  description: string;
  scentNotes: string[];
  category: string;
  priceCents: number;
  imageUrl: string;
  stock: number;
  featured: boolean;
  createdAt: Date;
}