# lib/api-zod/src/generated/types/productGender.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Scent Outlet perfume store API
 * OpenAPI spec version: 0.1.0
 */

export type ProductGender = typeof ProductGender[keyof typeof ProductGender];


export const ProductGender = {
  men: 'men',
  women: 'women',
  unisex: 'unisex',
} as const;